pyproject.toml
[project]
name = "ai-nova"
version = "0.1.0"
description = "AI-Nova repository management with Claude documentation workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "claude-code-sdk>=0.0.20",
    "boto3>=1.34.0",
    "awscli>=1.32.0",
]

[project.scripts]
pull-ai-nova = "pull_ai_nova:main"
